// test/resources/customTypes/enums.ts
export enum CategoryType {
    Clothing = "Clothing",
    Shoes = "Shoes",
    Furniture = "Furniture",
    Toys = "Toys",
    AudioSets = "Audio sets",
    Books = "Books"
}

export enum SectionType {
    NewArrivals = "New Arrivals",
    TrendingProducts = "Trending Products",
    TopRatedProducts = "Top-Rated Products",
    BestSellers = "Best Sellers"
}